requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.0
langchain==0.1.0
langchain-google-genai==1.0.0
python-dateutil==2.8.2
